Simplify directory-reading routines
Keep only Unix and Win32 versions and inline them.
